tests/wpt/tests/intersection-observer/intersection-ratio-with-fractional-bounds-2.html

<!DOCTYPE html>
<title>IntersectionObserver ratio with fractional bounds</title>
<link rel="help" href="https://bugs.chromium.org/p/chromium/issues/detail?id=1278897">
<meta name="viewport" content="width=device-width,initial-scale=1">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<style>
  #container {
    overflow: hidden;
    width: 500px;
  }
  #target {
    display: list-item;
    padding: 16.1px;
    background: rebeccapurple;
  }
</style>

<div id="container">
  <div id="list">
    <span id="target"></span>
  </div>
</div>

<script>
async_test(function(t) {
  let target = document.getElementById("target");
  let list = document.getElementById("list");
  let observer = new IntersectionObserver(t.step_func_done(function(entries) {
    assert_equals(entries.length, 1);
    assert_equals(entries[0].intersectionRatio, 1);
    assert_equals(entries[0].isIntersecting, true);
  }), { root: list, threshold: 1 });
  observer.observe(target);
});
</script>
